+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\alond\\Documents\\GitHub\\bookstore\\src\\index.js":"1","C:\\Users\\alond\\Documents\\GitHub\\bookstore\\src\\App.js":"2","C:\\Users\\alond\\Documents\\GitHub\\bookstore\\src\\components\\BookstoreFeatured.js":"3","C:\\Users\\alond\\Documents\\GitHub\\bookstore\\src\\components\\BookstoreItems.js":"4","C:\\Users\\alond\\Documents\\GitHub\\bookstore\\src\\components\\BooksList.js":"5","C:\\Users\\alond\\Documents\\GitHub\\bookstore\\src\\components\\BookItem.js":"6","C:\\Users\\alond\\Documents\\GitHub\\bookstore\\src\\components\\MoviesList.js":"7","C:\\Users\\alond\\Documents\\GitHub\\bookstore\\src\\components\\MovieItem.js":"8"},{"size":148,"mtime":1621043139363,"results":"9","hashOfConfig":"10"},{"size":1489,"mtime":1621806088594,"results":"11","hashOfConfig":"10"},{"size":340,"mtime":1621220473751,"results":"12","hashOfConfig":"10"},{"size":1035,"mtime":1621820961882,"results":"13","hashOfConfig":"10"},{"size":405,"mtime":1621818541170,"results":"14","hashOfConfig":"10"},{"size":288,"mtime":1621220366131,"results":"15","hashOfConfig":"10"},{"size":421,"mtime":1621806307072,"results":"16","hashOfConfig":"10"},{"size":260,"mtime":1621805837811,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},"w0m1wh",{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},"C:\\Users\\alond\\Documents\\GitHub\\bookstore\\src\\index.js",[],["35","36"],"C:\\Users\\alond\\Documents\\GitHub\\bookstore\\src\\App.js",[],"C:\\Users\\alond\\Documents\\GitHub\\bookstore\\src\\components\\BookstoreFeatured.js",[],"C:\\Users\\alond\\Documents\\GitHub\\bookstore\\src\\components\\BookstoreItems.js",[],"C:\\Users\\alond\\Documents\\GitHub\\bookstore\\src\\components\\BooksList.js",[],"C:\\Users\\alond\\Documents\\GitHub\\bookstore\\src\\components\\BookItem.js",[],"C:\\Users\\alond\\Documents\\GitHub\\bookstore\\src\\components\\MoviesList.js",[],"C:\\Users\\alond\\Documents\\GitHub\\bookstore\\src\\components\\MovieItem.js",[],{"ruleId":"37","replacedBy":"38"},{"ruleId":"39","replacedBy":"40"},"no-native-reassign",["41"],"no-negated-in-lhs",["42"],"no-global-assign","no-unsafe-negation"]
+[{"C:\\Users\\alond\\Documents\\GitHub\\bookstore\\src\\index.js":"1","C:\\Users\\alond\\Documents\\GitHub\\bookstore\\src\\App.js":"2","C:\\Users\\alond\\Documents\\GitHub\\bookstore\\src\\components\\Layout\\Header.js":"3","C:\\Users\\alond\\Documents\\GitHub\\bookstore\\src\\components\\Cart\\Cart.js":"4","C:\\Users\\alond\\Documents\\GitHub\\bookstore\\src\\Data\\Books_Data.js":"5","C:\\Users\\alond\\Documents\\GitHub\\bookstore\\src\\Data\\Movies_Data.js":"6","C:\\Users\\alond\\Documents\\GitHub\\bookstore\\src\\components\\Bookstore\\BookstoreFeatured.js":"7","C:\\Users\\alond\\Documents\\GitHub\\bookstore\\src\\components\\Bookstore\\BookstoreItems.js":"8","C:\\Users\\alond\\Documents\\GitHub\\bookstore\\src\\components\\Items\\Books\\BooksList.js":"9","C:\\Users\\alond\\Documents\\GitHub\\bookstore\\src\\components\\Items\\Movies\\MoviesList.js":"10","C:\\Users\\alond\\Documents\\GitHub\\bookstore\\src\\components\\Items\\Movies\\MovieItem.js":"11","C:\\Users\\alond\\Documents\\GitHub\\bookstore\\src\\components\\Items\\Books\\BookItem.js":"12"},{"size":148,"mtime":1621043139363,"results":"13","hashOfConfig":"14"},{"size":470,"mtime":1622260279891,"results":"15","hashOfConfig":"14"},{"size":228,"mtime":1622260680292,"results":"16","hashOfConfig":"14"},{"size":223,"mtime":1622259746979,"results":"17","hashOfConfig":"14"},{"size":909,"mtime":1622259798499,"results":"18","hashOfConfig":"14"},{"size":290,"mtime":1622259800990,"results":"19","hashOfConfig":"14"},{"size":371,"mtime":1622260344076,"results":"20","hashOfConfig":"14"},{"size":1063,"mtime":1622260560946,"results":"21","hashOfConfig":"14"},{"size":405,"mtime":1622258400029,"results":"22","hashOfConfig":"14"},{"size":421,"mtime":1622258407935,"results":"23","hashOfConfig":"14"},{"size":260,"mtime":1622258407846,"results":"24","hashOfConfig":"14"},{"size":256,"mtime":1622259732909,"results":"25","hashOfConfig":"14"},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},"w0m1wh",{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"31","messages":"32","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"35","messages":"36","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"37","messages":"38","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"39","messages":"40","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"41","messages":"42","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"43","messages":"44","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"45","messages":"46","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"47","messages":"48","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"49","messages":"50","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\alond\\Documents\\GitHub\\bookstore\\src\\index.js",[],["51","52"],"C:\\Users\\alond\\Documents\\GitHub\\bookstore\\src\\App.js",[],"C:\\Users\\alond\\Documents\\GitHub\\bookstore\\src\\components\\Layout\\Header.js",["53"],"C:\\Users\\alond\\Documents\\GitHub\\bookstore\\src\\components\\Cart\\Cart.js",[],"C:\\Users\\alond\\Documents\\GitHub\\bookstore\\src\\Data\\Books_Data.js",[],"C:\\Users\\alond\\Documents\\GitHub\\bookstore\\src\\Data\\Movies_Data.js",[],"C:\\Users\\alond\\Documents\\GitHub\\bookstore\\src\\components\\Bookstore\\BookstoreFeatured.js",[],"C:\\Users\\alond\\Documents\\GitHub\\bookstore\\src\\components\\Bookstore\\BookstoreItems.js",[],"C:\\Users\\alond\\Documents\\GitHub\\bookstore\\src\\components\\Items\\Books\\BooksList.js",[],"C:\\Users\\alond\\Documents\\GitHub\\bookstore\\src\\components\\Items\\Movies\\MoviesList.js",[],"C:\\Users\\alond\\Documents\\GitHub\\bookstore\\src\\components\\Items\\Movies\\MovieItem.js",[],"C:\\Users\\alond\\Documents\\GitHub\\bookstore\\src\\components\\Items\\Books\\BookItem.js",[],{"ruleId":"54","replacedBy":"55"},{"ruleId":"56","replacedBy":"57"},{"ruleId":"58","severity":1,"message":"59","line":1,"column":10,"nodeType":"60","messageId":"61","endLine":1,"endColumn":18},"no-native-reassign",["62"],"no-negated-in-lhs",["63"],"no-unused-vars","'Fragment' is defined but never used.","Identifier","unusedVar","no-global-assign","no-unsafe-negation"]
