+++ src/Data/Movies_Data.js
@@ -1,0 +1,12 @@
+const MOVIES = [
+  {
+    id: "m1",
+    title: "It's A Wonderful Life",
+    director: "Frank Capra",
+    year: 1946,
+    description:
+      "Consectetur duis pariatur deserunt ut aute in magna eu proident exercitation consectetur qui irure culpa.",
+  },
+];
+
+export default MOVIES;
