+++ src/components/Cart/Cart.js
@@ -1,0 +1,13 @@
+import { Fragment } from "react";
+
+const Cart = (props) => {
+  return (
+    <Fragment>
+      <div>CART ICON</div>
+      <div>BADGE</div>
+      <div>VIEW CART</div>
+    </Fragment>
+  );
+};
+
+export default Cart;
