+++ src/components/Layout/Header.js
@@ -1,0 +1,13 @@
+import { Fragment } from "react";
+import Cart from "../Cart/Cart";
+
+const Header = (props) => {
+  return (
+    <header>
+      <h2>Let's get started!</h2>
+      <Cart />
+    </header>
+  );
+};
+
+export default Header;
